# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is api for social media application",
        "title": "Social Media API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/api",
    "paths": {
...
        "/notification/finds": {
            "get": {
                "description": "authentication and find notifications with receiver id cursor pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "cursor uuid for notification id",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit for notifications cursor pagination",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SwaggerResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/notification.NotificationCursorPagination"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/notification/read-all": {
            "patch": {
                "description": "authentication and update read notifications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "parameters": [
                    {
                        "description": "update notification payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notification.UpdateNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SwaggerResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/notification.UpdatedNotification"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.SecureUser": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "profileBackgroundUrl": {
                    "type": "string"
                },
                "profileUrl": {
                    "type": "string"
                },
                "providerType": {
                    "$ref": "#/definitions/models.ProviderType"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.NotificationType": {
            "type": "string",
            "enum": [
                "POST",
                "SHARE",
                "COMMENT",
                "REPLY",
                "TAG",
                "LIKE",
                "FOLLOW"
            ],
            "x-enum-varnames": [
                "NotificationTypePost",
                "NotificationTypeShare",
                "NotificationTypeComment",
                "NotificationTypeReply",
                "NotificationTypeTag",
                "NotificationTypeLike",
                "NotificationTypeFollow"
            ]
        },
        "models.ProviderType": {
            "type": "string",
            "enum": [
                "LOCAL",
                "GOOGLE",
                "FACEBOOK",
                "GITHUB"
            ],
            "x-enum-varnames": [
                "ProviderTypeLocal",
                "ProviderTypeGoogle",
                "ProviderTypeFacebook",
                "ProviderTypeGithub"
            ]
        },
        "models.Role": {
            "type": "string",
            "enum": [
                "USER",
                "ADMIN"
            ],
            "x-enum-varnames": [
                "RoleUser",
                "RoleAdmin"
            ]
        },
        "notification.Notification": {
            "type": "object",
            "properties": {
                "commentId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isRead": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "postId": {
                    "type": "string"
                },
                "receiverId": {
                    "type": "string"
                },
                "sender": {
                    "$ref": "#/definitions/dto.SecureUser"
                },
                "senderId": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.NotificationType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "notification.NotificationCursorPagination": {
            "type": "object",
            "properties": {
                "nextCursor": {
                    "type": "string"
                },
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/notification.Notification"
                    }
                }
            }
        },
        "notification.UpdateNotificationRequest": {
            "type": "object",
            "required": [
                "notificationsId"
            ],
            "properties": {
                "notificationsId": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "notification.UpdatedNotification": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "isRead": {
                    "type": "boolean"
                }
            }
        },
...
        "response.SwaggerBadRequestResponse": {
            "type": "object",
            "properties": {
                "errorFields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.SwaggerResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.SwaggerResponseWithData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
...
}